# [opt](hive)opt select count(*) stmt push down agg on parquet in hive .