--- a/be/src/vec/exec/format/generic_reader.h
+++ b/be/src/vec/exec/format/generic_reader.h
@@ -30,12 +30,11 @@ class Block;
 // a set of blocks with specified schema,
 class GenericReader {
 public:
+    GenericReader() : _push_down_agg_type(TPushAggOp::type::NONE) {}
+    void set_push_down_agg_type(TPushAggOp::type push_down_agg_type) {
+        _push_down_agg_type = push_down_agg_type;
+    }
     virtual Status get_next_block(Block* block, size_t* read_rows, bool* eof) = 0;
-
-    virtual Status get_next_block(Block* block, size_t* read_rows, bool* eof,
-                                  TPushAggOp::type push_down_agg_type_opt) {
-        return Status::NotSupported("not support this type!");
-    };
 
     virtual std::unordered_map<std::string, TypeDescriptor> get_name_to_type() {
         std::unordered_map<std::string, TypeDescriptor> map;
@@ -71,6 +70,7 @@ protected:
 
     /// Whether the underlying FileReader has filled the partition&missing columns
     bool _fill_all_columns = false;
+    TPushAggOp::type _push_down_agg_type;
 };
 
 } // namespace doris::vectorized