--- a/be/src/vec/exec/scan/new_olap_scan_node.cpp
+++ b/be/src/vec/exec/scan/new_olap_scan_node.cpp
@@ -249,9 +249,8 @@ Status NewOlapScanNode::_process_conjuncts() {
 }
 
 Status NewOlapScanNode::_build_key_ranges_and_filters() {
-    if (!_olap_scan_node.__isset.push_down_agg_type_opt ||
-        _olap_scan_node.push_down_agg_type_opt == TPushAggOp::NONE) {
-        const std::vector<std::string>& column_names = _olap_scan_node.key_column_name;
+    if (push_down_agg_type_opt == TPushAggOp::NONE) {
+ 	const std::vector<std::string>& column_names = _olap_scan_node.key_column_name;
         const std::vector<TPrimitiveType::type>& column_types = _olap_scan_node.key_column_type;
         DCHECK(column_types.size() == column_names.size());
 
@@ -328,7 +327,7 @@ Status NewOlapScanNode::_build_key_ranges_and_filters() {
     } else {
         _runtime_profile->add_info_string(
                 "PushDownAggregate",
-                push_down_agg_to_string(_olap_scan_node.push_down_agg_type_opt));
+    		push_down_agg_to_string(push_down_agg_type_opt));
     }
 
     if (_state->enable_profile()) {