--- a/be/src/vec/exec/scan/new_olap_scanner.cpp
+++ b/be/src/vec/exec/scan/new_olap_scanner.cpp
@@ -266,7 +266,7 @@ Status NewOlapScanner::_init_tablet_reader_params(
         _aggregation = true;
     } else {
         _tablet_reader_params.direct_mode = _aggregation || single_version ||
-                                            (_parent->push_down_agg_type_opt != TPushAggOp::NONE);
+                                            (_parent->get_push_down_agg_type() != TPushAggOp::NONE);
     }
 
     RETURN_IF_ERROR(_init_return_columns());
@@ -275,9 +275,7 @@ Status NewOlapScanner::_init_tablet_reader_params(
     _tablet_reader_params.tablet_schema = _tablet_schema;
     _tablet_reader_params.reader_type = ReaderType::READER_QUERY;
     _tablet_reader_params.aggregation = _aggregation;
-    if (_parent->push_down_agg_type_opt) {
-        _tablet_reader_params.push_down_agg_type_opt = _parent->push_down_agg_type_opt;
-    }
+    _tablet_reader_params.push_down_agg_type_opt = _parent->get_push_down_agg_type();
     _tablet_reader_params.version = Version(0, _version);
 
     // TODO: If a new runtime filter arrives after `_conjuncts` move to `_common_expr_ctxs_push_down`,