--- a/be/src/vec/exec/scan/scanner_scheduler.cpp
+++ b/be/src/vec/exec/scan/scanner_scheduler.cpp
@@ -49,6 +49,7 @@
 #include "vec/exec/scan/new_olap_scanner.h" // IWYU pragma: keep
 #include "vec/exec/scan/scanner_context.h"
 #include "vec/exec/scan/vscanner.h"
+#include "vec/exec/scan/vscan_node.h"
 #include "vfile_scanner.h"
 
 namespace doris::vectorized {