… fe/fe-core/src/main/java/org/apache/doris/nereids/rules/implementation/AggregateStrategies.java
… fe/fe-core/src/main/java/org/apache/doris/planner/OlapScanNode.java
… fe/fe-core/src/main/java/org/apache/doris/planner/PlanNode.java
… fe/fe-core/src/main/java/org/apache/doris/planner/SingleNodePlanner.java
… fe/fe-core/src/main/java/org/apache/doris/planner/external/FileScanNode.java
… fe/fe-core/src/main/java/org/apache/doris/planner/external/HiveScanNode.java
… gensrc/thrift/PlanNodes.thrift
… regression-test/conf/regression-conf.groovy
… regression-test/data/external_table_emr_p2/hive/test_select_count_optimize.out
… regression-test/data/performance_p0/redundant_conjuncts.out
… regression-test/suites/external_table_emr_p2/hive/test_select_count_optimize.groovy