## Proposed changes

after pr :  #14827 and #12803 

<!--Describe your changes.-->
Optimization "select count(*) from table" stmtement , push down "count" type to <code>be</code> .

support file type : parquet in hive .


1. 4kfiles , 60kwline num 

    before:  1 min 37.70 sec 

    after:   50.18 sec



2. 50files , 60kwline num

    before: 1.12 sec

    after: 0.82 sec



## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...