@morningman 2023-07-27
run buildall